mostrando pie chart com porcentagens dos status dos desejos

+++ lib/data/models/statistic_percentage_status_comparison.dart
@@ -1,12 +1,12 @@
 
 class StatisticPercentageStatusComparison {
-  int numberOfAccomplished;
-  int numberOfAccomplishedInAdvance;
-  int numberOfPending;
-  int numberOfNotAccomplishedUntilTargetDate;
+  double numberOfAccomplished;
+  double numberOfAccomplishedInAdvance;
+  double numberOfPending;
+  double numberOfNotAccomplishedUntilTargetDate;
 
   StatisticPercentageStatusComparison({required this.numberOfAccomplished,
         required this.numberOfAccomplishedInAdvance, required this.numberOfPending,
         required this.numberOfNotAccomplishedUntilTargetDate
-        });
+  });
 }
